majdnem

+++ school_db/database/migrations/2025_03_18_102202_create_students_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
 
-            $table->unsignedInteger('class_id');
+            $table->unsignedBigInteger('class_id');
             $table->foreign('class_id')->references('id')->on('school_classes');
 
             $table->string('gender');
